Split VictoryProgressUpdater to VictoryProgressUpdater and VictoryTypesProcessorScriptDataAcessor.
