+++ sources/Colonization/Activities/HostActivity/HostActivity.cpp
@@ -22,7 +22,7 @@
 #include <Colonization/Backend/PlayersPointsCalculator.hpp>
 #include <Colonization/Backend/TradeProcessor.hpp>
 #include <Colonization/Backend/UnitsManager.hpp>
-#include <Colonization/Backend/VictoryProgressUpdater.hpp>
+#include <Colonization/Backend/VictoryProgressUpdater/VictoryProgressUpdater.hpp>
 
 namespace Colonization
 {
