+++ sources/Colonization/AngelScriptBinders/Backend/BindVictoryTypesProcessorScriptDataAccessor.cpp
@@ -2,7 +2,7 @@
 #include "BindVictoryTypesProcessorScriptDataAccessor.hpp"
 #include <Urho3D/ThirdParty/AngelScript/angelscript.h>
 #include <Urho3D/AngelScript/APITemplates.h>
-#include <Colonization/Backend/VictoryProgressUpdater.hpp>
+#include <Colonization/Backend/VictoryProgressUpdater/VictoryTypesProcessorScriptDataAccessor.hpp>
 
 namespace Colonization
 {
@@ -16,6 +16,7 @@ void BindVictoryTypesProcessorScriptDataAccessor (Urho3D::Script *script)
 void BindVictoryTypesProcessorScriptDataAccessorInterface (Urho3D::Script *script, Urho3D::String className)
 {
     asIScriptEngine *engine = script->GetScriptEngine ();
+    engine->RegisterObjectMethod (className.CString (), "const GameConfiguration @+ get_gameConfiguration () const", asMETHOD (VictoryTypesProcessorScriptDataAccessor, GetGameConfiguration), asCALL_THISCALL);
     engine->RegisterObjectMethod (className.CString (), "const PlayerInfo @+ get_playerInfo () const", asMETHOD (VictoryTypesProcessorScriptDataAccessor, GetPlayerInfo), asCALL_THISCALL);
     engine->RegisterObjectMethod (className.CString (), "uint get_districtsCount () const", asMETHOD (VictoryTypesProcessorScriptDataAccessor, GetDistrictsCount), asCALL_THISCALL);
     engine->RegisterObjectMethod (className.CString (), "const District @+ GetDistrictByIndex (int index) const", asMETHOD (VictoryTypesProcessorScriptDataAccessor, GetDistrictByIndex), asCALL_THISCALL);
