+++ sources/Colonization/Backend/VictoryProgressUpdater/VictoryTypesProcessorScriptDataAccessor.hpp
@@ -12,38 +12,6 @@
 
 namespace Colonization
 {
-// TODO: Maybe split to VictoryProgressUpdater, VictoryTypesProcessorScriptDataAccessor.
-class VictoryProgressUpdater : public Urho3D::Component
-{
-URHO3D_OBJECT (VictoryProgressUpdater, Component)
-protected:
-    Urho3D::SharedPtr <Urho3D::ScriptFile> victoryTypesProcessor_;
-    float timeUntilGameEnd_;
-    bool isAnyoneWon_;
-    Urho3D::String winnerName_;
-    Urho3D::String victoryType_;
-    Urho3D::String victoryInfo_;
-
-    void UpdateVictoryByPointsProgresses ();
-    void SetWinnerFromVictoryByPoints ();
-    void CheckForAnyVictory ();
-    void ProcessScriptedVictoryTypes (float timeStep);
-    virtual void OnSceneSet (Urho3D::Scene* scene);
-
-public:
-    VictoryProgressUpdater (Urho3D::Context *context);
-    virtual ~VictoryProgressUpdater ();
-
-    static void RegisterObject (Urho3D::Context *context);
-    void Update (Urho3D::StringHash eventType, Urho3D::VariantMap &eventData);
-
-    float GetTimeUntilGameEnd ();
-    bool IsAnyoneWon ();
-    Urho3D::String GetWinnerName ();
-    Urho3D::String GetVictoryType ();
-    Urho3D::String GetVictoryInfo ();
-};
-
 class VictoryTypesProcessorScriptDataAccessor : public Urho3D::Object
 {
 URHO3D_OBJECT (VictoryTypesProcessorScriptDataAccessor, Object)
@@ -52,12 +20,15 @@ protected:
     UnitsManager *unitsManager_;
     TradeProcessor *tradeProcessor_;
     PlayerInfo *playerInfo_;
+    GameConfiguration *configuration_;
 
 public:
     VictoryTypesProcessorScriptDataAccessor (Urho3D::Context *context);
     virtual ~VictoryTypesProcessorScriptDataAccessor ();
 
-    void Setup (Map *map, UnitsManager *unitsManager, TradeProcessor *tradeProcessor);
+    void Setup (Map *map, UnitsManager *unitsManager, TradeProcessor *tradeProcessor, GameConfiguration *configuration);
+    const GameConfiguration *GetGameConfiguration () const;
+
     void SetPlayerInfo (PlayerInfo *playerInfo);
     const PlayerInfo *GetPlayerInfo () const;
